+++ Assets/Adjust/Android/AdjustAndroid.cs
@@ -8,7 +8,7 @@ namespace com.adjust.sdk
 #if UNITY_ANDROID
     public class AdjustAndroid
     {
-        private const string sdkPrefix = "unity4.23.2";
+        private const string sdkPrefix = "unity4.24.0";
         private static bool launchDeferredDeeplink = true;
         private static AndroidJavaClass ajcAdjust = new AndroidJavaClass("com.adjust.sdk.Adjust");
         private static AndroidJavaObject ajoCurrentActivity = new AndroidJavaClass("com.unity3d.player.UnityPlayer").GetStatic<AndroidJavaObject>("currentActivity");
@@ -84,6 +84,18 @@ namespace com.adjust.sdk
             if (adjustConfig.sendInBackground != null)
             {
                 ajoAdjustConfig.Call("setSendInBackground", adjustConfig.sendInBackground.Value);
+            }
+
+            // Check if user wants to get cost data in attribution callback.
+            if (adjustConfig.needsCost != null)
+            {
+                ajoAdjustConfig.Call("setNeedsCost", adjustConfig.needsCost.Value);
+            }
+
+            // Check if user wants to run preinstall campaigns.
+            if (adjustConfig.preinstallTrackingEnabled != null)
+            {
+                ajoAdjustConfig.Call("setPreinstallTrackingEnabled", adjustConfig.preinstallTrackingEnabled.Value);
             }
 
             // Check if user has set user agent value.
@@ -311,6 +323,21 @@ namespace com.adjust.sdk
                     null : ajoAttribution.Get<string>(AdjustUtils.KeyClickLabel);
                 adjustAttribution.adid = ajoAttribution.Get<string>(AdjustUtils.KeyAdid) == "" ?
                     null : ajoAttribution.Get<string>(AdjustUtils.KeyAdid);
+                adjustAttribution.costType = ajoAttribution.Get<string>(AdjustUtils.KeyCostType) == "" ?
+                    null : ajoAttribution.Get<string>(AdjustUtils.KeyCostType);
+                AndroidJavaObject ajoCostAmount = ajoAttribution.Get<AndroidJavaObject>(AdjustUtils.KeyCostAmount) == null ?
+                    null : ajoAttribution.Get<AndroidJavaObject>(AdjustUtils.KeyCostAmount);
+                if (ajoCostAmount == null)
+                {
+                    adjustAttribution.costAmount = null;
+                }
+                else
+                {
+                    double costAmount = ajoCostAmount.Call<double>("doubleValue");
+                    adjustAttribution.costAmount = costAmount;
+                }
+                adjustAttribution.costCurrency = ajoAttribution.Get<string>(AdjustUtils.KeyCostCurrency) == "" ?
+                    null : ajoAttribution.Get<string>(AdjustUtils.KeyCostCurrency);
                 return adjustAttribution;
             }
             catch (Exception) {}
@@ -506,6 +533,21 @@ namespace com.adjust.sdk
                     null : attribution.Get<string>(AdjustUtils.KeyClickLabel);
                 adjustAttribution.adid = attribution.Get<string>(AdjustUtils.KeyAdid) == "" ?
                     null : attribution.Get<string>(AdjustUtils.KeyAdid);
+                adjustAttribution.costType = attribution.Get<string>(AdjustUtils.KeyCostType) == "" ?
+                    null : attribution.Get<string>(AdjustUtils.KeyCostType);
+                AndroidJavaObject ajoCostAmount = attribution.Get<AndroidJavaObject>(AdjustUtils.KeyCostAmount) == null ?
+                    null : attribution.Get<AndroidJavaObject>(AdjustUtils.KeyCostAmount);
+                if (ajoCostAmount == null)
+                {
+                    adjustAttribution.costAmount = null;
+                }
+                else
+                {
+                    double costAmount = ajoCostAmount.Call<double>("doubleValue");
+                    adjustAttribution.costAmount = costAmount;
+                }
+                adjustAttribution.costCurrency = attribution.Get<string>(AdjustUtils.KeyCostCurrency) == "" ?
+                    null : attribution.Get<string>(AdjustUtils.KeyCostCurrency);
                 callback(adjustAttribution);
             }
         }